Trait arc::geom::rt::trace::Trace

pub trait Trace {
    fn hit(&self, ray: &Ray) -> bool;

    fn dist(&self, ray: &Ray) -> Option<f64>;

    fn dist_norm(&self, ray: &Ray) -> Option<(f64, Unit<Vector3<f64>>)>;

    fn dist_inside(&self, ray: &Ray) -> Option<(f64, bool)>;

    fn dist_inside_norm(
        &self, 
        ray: &Ray
    ) -> Option<(f64, bool, Unit<Vector3<f64>>)>;

#[must_use]
    fn hit_point(&self, ray: &Ray) -> Option<Point3<f64>> { ... }
}

Trace trait implementation. Types implementing this trait can be traced using rays.

Required methods

fn hit(&self, ray: &Ray) -> bool

Determine if a ray hit occurs.

fn dist(&self, ray: &Ray) -> Option<f64>

Distance to the surface along the ray's line of travel.

fn dist_norm(&self, ray: &Ray) -> Option<(f64, Unit<Vector3<f64>>)>

Distance to the surface along the ray's line of travel and normal unit vector at the point of collision.

fn dist_inside(&self, ray: &Ray) -> Option<(f64, bool)>

Distance to the surface along the ray's line of travel and side of collision.

fn dist_inside_norm(&self, ray: &Ray) -> Option<(f64, bool, Unit<Vector3<f64>>)>

Distance to the surface along the ray's line of travel, side of collision, and normal unit vector at the point of collision.

Loading content...
Provided methods

#[must_use] fn hit_point(&self, ray: &Ray) -> Option<Point3<f64>>

Calculate the hit point of a ray.

Loading content...