Type Definition arc::ord::set::inter_set::InterSet

type InterSet = Set<InterKey, Interface>;

Alias for the interface set.

Methods

impl InterSet

#[must_use] pub fn surf_keys(&self) -> Vec<SurfKey>

Get a list of all surface keys used by the interface set.

#[must_use] pub fn mat_keys(&self) -> Vec<MatKey>

Get a list of all material keys used by the interface set.

#[must_use] pub fn observe_mat(
&self,
surfs: &SurfSet,
bound: &Aabb,
ray: &Ray
) -> Option<&MatKey>

Determine which material, if any, would be observed with a given ray.