Type Definition arc::ord::set::region_set::RegionSet

type RegionSet = Set<RegionKey, Region>;

Alias for the region set.

Methods

impl RegionSet

#[must_use] pub fn surf_keys(&self) -> Vec<SurfKey>

Get a list of all surface keys used by the interface set.

#[must_use] pub fn state_keys(&self) -> Vec<StateKey>

Get a list of all state keys used by the interface set.

#[must_use] pub fn observe_state(
&self,
surfs: &SurfSet,
bound: &Aabb,
ray: &Ray
) -> Option<StateKey>

Determine which state, if any, would be observed with a given ray.