Struct arc::ord::set::Set

pub struct Set<K: Ord, T> { /* fields omitted */ }

Set mapping.

Methods

impl Set<InterKey, Interface>

#[must_use] pub fn surf_keys(&self) -> Vec<SurfKey>

Get a list of all surface keys used by the interface set.

#[must_use] pub fn mat_keys(&self) -> Vec<MatKey>

Get a list of all material keys used by the interface set.

#[must_use] pub fn observe_mat(
&self,
surfs: &SurfSet,
bound: &Aabb,
ray: &Ray
) -> Option<&MatKey>

Determine which material, if any, would be observed with a given ray.

impl Set<LightKey, Light>

#[must_use] pub fn surf_keys(&self) -> Vec<SurfKey>

Get a list of all surface keys used by the light set.

impl Set<ReactKey, Reaction>

#[must_use] pub fn spec_keys(&self) -> Vec<SpecKey>

Get a list of all species keys used by the reaction set.

impl Set<RegionKey, Region>

#[must_use] pub fn surf_keys(&self) -> Vec<SurfKey>

Get a list of all surface keys used by the interface set.

#[must_use] pub fn state_keys(&self) -> Vec<StateKey>

Get a list of all state keys used by the interface set.

#[must_use] pub fn observe_state(
&self,
surfs: &SurfSet,
bound: &Aabb,
ray: &Ray
) -> Option<&StateKey>

Determine which state, if any, would be observed with a given ray.

impl Set<StateKey, State>

#[must_use] pub fn spec_keys(&self) -> Vec<SpecKey>

Get a list of all species keys used by the state set.

impl Set<SurfKey, Mesh>

#[must_use] pub fn build(
proto_surfs: &Set<SurfKey, SurfaceForm>,
mesh_set: &MeshSet
) -> Self

Form a surface set from a mesh set.

impl<K: Display + Ord, T> Set<K, T>

#[must_use] pub fn map(&self) -> &BTreeMap<K, T>

#[must_use] pub fn new(map: BTreeMap<K, T>) -> Self

Construct a new instance.

#[must_use] pub fn get(&self, key: &K) -> &T

Access a value within the map.

#[must_use] pub fn index_of_key(&self, key: &K) -> usize

Determine the index corresponding to a key within the set.

impl<K: Display + Clone + Ord, T: Load> Set<K, T>

#[must_use] pub fn load(dir: &Path, keys: &[K], ext: &str) -> Self

Load a set of files.