Crate arc

Library core.

Modules

chem
Chemistry module.

data
Data module.

file
File module.

fmt
Formatting module.

geom
Geometry module.

img
Rendering module.

list
Lists module.

math
Mathematics module.

ord
Ordering module.

phys
Physics module.

sim
Simulation module.

util
Utility module.

world
World module.

Macros

access
Succinctly create a reference method for the given variable.

args
Import command line arguments as a requested type.

clone
Succinctly create a clone method for the given variable.

columns
Report a list of values in evenly spaced columns.

map
Concisely construct a hashmap.

report
Report a value and either its associated name, or a human readable string if supplied.

rows
Report a row of values in equally spaced columns.

values
Report a list of values in evenly spaced columns.